+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - renamed IFD0 in the output object from `output.image` to `output.ifd0`
 - renamed `thumbnailBuffer()` function to `thumbnail()`. It now also returns `Uint8Array` instead of `ArrayBuffer` in browser. Node.js version keeps returning `Buffer`.
 - renamed & simplified behavior of `seekChunkSize` and `parseChunkSize`. See `firstChunkSize`, `firstChunkSizeBrowser`, `firstChunkSizeNode`.
-- removed `postProcess` property and split its behavior to new properties `sanitize`, `translateTags`, `translateValues` and `reviveValues`.
+- removed `postProcess` property and split its behavior to new properties `sanitize`, `translateKeys`, `translateValues` and `reviveValues`.
 
 
 ### Added